requirements.txt
aiohttp==3.8.3
beautifulsoup4==4.11.1
deepl_translate==1.2.0
pubchempy==1.0.4
playwright==1.27.1
py-cord==2.2.2
python-dotenv==0.21.0